display with emotions/requirements.txt
colorzero==2.0
gpiozero==2.0.1
numpy==2.2.4
pillow==11.1.0
RPi.GPIO==0.7.1
spi==0.2.0
spidev==3.6
